Color system fixed bullet system fixed

+++ Source/Engine/Renderer/Renderer.cpp
@@ -37,7 +37,7 @@ namespace Jackster {
 		SDL_RenderPresent(m_renderer);
 	}
 
-	void Renderer::setColor(int R, int G, int B, int A)
+	void Renderer::setColor(uint8_t R, uint8_t G, uint8_t B, uint8_t A)
 	{
 		SDL_SetRenderDrawColor(m_renderer, R, G, B, A);
 	}
